requirements.txt
altair==5.4.1
attrs==24.2.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
kaggle==1.6.17
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
narwhals==1.10.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
plotly==5.24.1
protobuf==5.28.3
py4j==0.10.9.7
pyarrow==17.0.0
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.4
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
webencodings==0.5.1
zipp==3.20.2
